!!! BUG !!!
charts: mlperf-client-monitor: Specify the port name in ServiceImport

The name of the port of the Service import must be defined. Currently
the server has a Service with port name default. This name must be
specified in the ServiceImport in the client. If not, it will not work
(connection refused when request the corresponding URL). In a future commit
this port name will be added in the values.yml in both server and client.

Signed-off-by: Orfeasfil <orfeasfil2000@gmail.com>